+++ trunk/Commun/Utilitaire/AideGL.cpp
@@ -181,7 +181,7 @@ namespace aidegl {
       GLint Cloture[4];
       glGetIntegerv( GL_VIEWPORT, Cloture);
       glLoadIdentity() ;
-      gluOrtho2D( 0.0, (GLdouble) Cloture[2], (GLdouble) Cloture[3], 0.0 );
+      gluOrtho2D( 0.0, (GLfloat) Cloture[2], (GLfloat) Cloture[3], 0.0 );
 
       // On trace le rectangle en pointillés.
       glEnable( GL_LINE_STIPPLE );
@@ -295,7 +295,7 @@ namespace aidegl {
 	   GLint Cloture[4];
 	   glGetIntegerv( GL_VIEWPORT, Cloture);
 	   glLoadIdentity() ;
-	   gluOrtho2D( 0.0, (GLdouble) Cloture[2], (GLdouble) Cloture[3], 0.0 );
+	   gluOrtho2D( 0.0, (GLfloat) Cloture[2], (GLfloat) Cloture[3], 0.0 );
 
 	   // On trace le rectangle en pointillés.
 	   glEnable( GL_LINE_STIPPLE );
