+++ trunk/Commun/Utilitaire/CLectureFichierBinaire.h
@@ -14,7 +14,7 @@
 /// @class CLectureFichierBinaire
 /// @brief cette classe contient des méthodes > permettant de lire
 ///        dans un fichier binaire des variables de types string,
-///        double, float, int, unsigned int, char, et bool.
+///        float, float, int, unsigned int, char, et bool.
 ///
 /// @author DGI-2990
 /// @date 2005-10-15
@@ -31,8 +31,8 @@ public:
 
    /// Surcharge de l'opérateur pour le type @em std::string.
    friend CLectureFichierBinaire& operator> ( CLectureFichierBinaire& in, std::string& s );
-   /// Surcharge de l'opérateur pour le type @em double.
-   friend CLectureFichierBinaire& operator> ( CLectureFichierBinaire& in, double& f );
+   /// Surcharge de l'opérateur pour le type @em float.
+   friend CLectureFichierBinaire& operator> ( CLectureFichierBinaire& in, float& f );
    /// Surcharge de l'opérateur pour le type @em float.
    friend CLectureFichierBinaire& operator> ( CLectureFichierBinaire& in, float& f );
    /// Surcharge de l'opérateur pour le type @em int.
